+++ Sources/Views/ALKMessageCell.swift
@@ -169,7 +169,7 @@ open class ALKFriendMessageCell: ALKMessageCell {
         nameLabel.setStyle(ALKMessageStyle.displayName)
         messageView.setStyle(ALKMessageStyle.receivedMessage)
         if ALKMessageStyle.receivedBubble.style == .edge {
-            bubbleView.tintColor = UIColor(netHex: 0xF1F0F0)
+            bubbleView.tintColor = ALKMessageStyle.receivedBubble.color
             bubbleView.image = bubbleViewImage(for: ALKMessageStyle.receivedBubble.style, isReceiverSide: true,showHangOverImage: false)
         }else{
             bubbleView.layer.cornerRadius = ALKMessageStyle.receivedBubble.cornerRadius
@@ -405,7 +405,7 @@ open class ALKMyMessageCell: ALKMessageCell {
         super.setupStyle()
         messageView.setStyle(ALKMessageStyle.sentMessage)
         if ALKMessageStyle.sentBubble.style == .edge {
-            bubbleView.tintColor = UIColor(netHex: 0xF1F0F0)
+            bubbleView.tintColor = ALKMessageStyle.sentBubble.color
             bubbleView.image = bubbleViewImage(for: ALKMessageStyle.sentBubble.style, isReceiverSide: false,showHangOverImage: false)
         }else{
             bubbleView.layer.cornerRadius = ALKMessageStyle.sentBubble.cornerRadius