+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The changelog for [ApplozicSwift](https://github.com/AppLozic/ApplozicSwift). Al
 - Fixed an issue where conversation details weren't getting refreshed when chat was opened from tapping on notification.
 - [AL-3307] Fixed an issue where typing indicator in 1-1 chat when user was typing in group.
 - [AL-3312] Fixed an issue where profile image of receiver wasn't visible for video cells.
+-  Fixed an issue where video weren't being send after recording from the SDK.
 
 
 2.3.0